Remade everything into uint cause that's really all we need.  Also, split the two interfaces since they're different enough to not have to reuse interface.

+++ rere-fencer/rere-fencer/Input/IGenomeReader.cs
@@ -18,12 +18,12 @@ namespace rere_fencer.Input
     public interface IGenomeContig
     {
         string Name { get; }
-        ulong Length { get; }
+        uint Length { get; }
         bool ContainsNs { get; }
         bool ContainsMaskedSequences { get; }
-        string GetSequence(ulong start, ulong end, bool ignoreMasks = false, bool ignoreNs = false);
-        char GetNucleotideAt(ulong position);
-        IEnumerable<char> GetNucleotides(ulong start, ulong end, bool ignoreMasks = false, bool ignoreNs = false); 
+        string GetSequence(uint start, uint end, bool ignoreMasks = false, bool ignoreNs = false);
+        char GetNucleotideAt(uint position);
+        IEnumerable<char> GetNucleotides(uint start, uint end, bool ignoreMasks = false, bool ignoreNs = false); 
         //uint? FirstPositionOfSequence(string sequence, uint offset = 0, bool strict = true);
         //uint? LastPositionOfSequence(string sequence, uint offset = 0, bool strict = true);
     }
